Last bits of pipeline reorg clean-up: removed some overkill all_specs_test.go files, audit the docs, updated changelog, moved all plugin registration to close to the plugin implementation, caught a test suite that wasn't being run.
