+++ CHANGES.txt
@@ -4,6 +4,9 @@
 Backwards Incompatibilities
 ---------------------------
 
+* Major reorganization of the `pipeline` package, moving the implementation of
+  most plugins to sub-packages of a separate `plugins` package.
+
 * Removed the wrapper 'table' element from the JSON serialization (issue #525)
   i.e., {"table":{"value":1}} is now simply {"value":1}.  The change also 
   removes the special '_name' metadata arribute; the top level _name element
